# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cost Analysis Per Container API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /containers:
    get:
      operationId: listContainers
      tags: [containers]
      summary: List all containers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of containers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerListResponse"
    post:
      operationId: createContainer
      tags: [containers]
      summary: Create a single container record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContainerRequest"
      responses:
        "201":
          description: Container created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/upload:
    post:
      operationId: uploadContainers
      tags: [containers]
      summary: Upload containers from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadContainersRequest"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadContainersResponse"
  /containers/{id}:
    get:
      operationId: getContainer
      tags: [containers]
      summary: Get a container by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Container details with charges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerDetail"
    put:
      operationId: updateContainer
      tags: [containers]
      summary: Update container basic info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContainerRequest"
      responses:
        "200":
          description: Updated container
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}/lock:
    post:
      operationId: lockContainer
      tags: [containers]
      summary: Lock or unlock a container record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockContainerRequest"
      responses:
        "200":
          description: Lock status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}/charges:
    get:
      operationId: getContainerCharges
      tags: [charges]
      summary: Get all charges for a container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerCharges"
    put:
      operationId: updateContainerCharges
      tags: [charges]
      summary: Update charges for a container section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContainerChargesRequest"
      responses:
        "200":
          description: Updated charges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerCharges"
  /containers/{id}/audit:
    get:
      operationId: getContainerAuditLog
      tags: [containers]
      summary: Get audit log for a container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
...
components:
  schemas:
...
    Container:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        containerNumber:
          type: string
        blNumber:
          type: string
        declaration:
          type: string
        size:
          type: string
        vessel:
          type: string
        status:
          type: string
          enum: [new_upload, documentation_review, shipping_entry, customs_entry, terminal_entry, delivery_entry, accounting_review, management_approval, completed, closed]
        isLocked:
          type: boolean
        assignedStaffId:
          type: integer
          nullable: true
        assignedStaffName:
          type: string
          nullable: true
        totalCost:
          type: number
        clearingCharges:
          type: number
        grossProfit:
          type: number
        dutyNotPaid:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, containerNumber, blNumber, declaration, size, vessel, status, isLocked, totalCost, clearingCharges, grossProfit, dutyNotPaid, createdAt, updatedAt]
    ContainerListResponse:
      type: object
      properties:
        containers:
          type: array
          items:
            $ref: "#/components/schemas/Container"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [containers, total, page, limit]
    CreateContainerRequest:
      type: object
      properties:
        customerName:
          type: string
        containerNumber:
          type: string
        blNumber:
          type: string
        declaration:
          type: string
        size:
          type: string
        vessel:
          type: string
        clearingCharges:
          type: number
      required: [customerName, containerNumber, blNumber, declaration, size, vessel]
    UpdateContainerRequest:
      type: object
      properties:
        customerName:
          type: string
        containerNumber:
          type: string
        blNumber:
          type: string
        declaration:
          type: string
        size:
          type: string
        vessel:
          type: string
        status:
          type: string
        assignedStaffId:
          type: integer
          nullable: true
        clearingCharges:
          type: number
    UploadContainersRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/UploadRow"
      required: [rows]
    UploadRow:
      type: object
      properties:
        customerName:
          type: string
        containerNumber:
          type: string
        blNumber:
          type: string
        declaration:
          type: string
        size:
          type: string
        vessel:
          type: string
        clearingCharges:
          type: number
      required: [customerName, containerNumber, blNumber]
    UploadContainersResponse:
      type: object
      properties:
        created:
          type: integer
        duplicates:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
      required: [created, duplicates, errors]
    LockContainerRequest:
      type: object
      properties:
        locked:
          type: boolean
        reason:
          type: string
      required: [locked]
    ContainerCharges:
      type: object
      properties:
        containerId:
          type: integer
        shipping:
          $ref: "#/components/schemas/ShippingCharges"
        customs:
          $ref: "#/components/schemas/CustomsCharges"
        terminal:
          $ref: "#/components/schemas/TerminalCharges"
        delivery:
          $ref: "#/components/schemas/DeliveryCharges"
        operations:
          $ref: "#/components/schemas/OperationsCharges"
        totalCost:
          type: number
        clearingCharges:
          type: number
        grossProfit:
          type: number
      required: [containerId, shipping, customs, terminal, delivery, operations, totalCost, clearingCharges, grossProfit]
    ShippingCharges:
      type: object
      properties:
        shippingCompany:
          type: number
        shippingPaymentVat:
          type: number
        consignee:
          type: number
        finalInvoiceShippingCompany:
          type: number
        telexCharge:
          type: number
        shippingRunnings:
          type: number
        shippingDetentionToBePaidByCustomer:
          type: number
    CustomsCharges:
      type: object
      properties:
        duty:
          type: number
        dutyPaid:
          type: number
        dutyNotPaid:
          type: number
        valuation:
          type: number
        ciu:
          type: number
        upCountryCustom:
          type: number
        dciu:
          type: number
        mdReleasingPackage:
          type: number
        ocSettlement:
          type: number
        ocReleaseLocal:
          type: number
        dcEnforcementForTransire:
          type: number
        complianceTeam:
          type: number
        cacSettlement:
          type: number
        crffn:
          type: number
        soncap:
          type: number
        alerts:
          type: number
        examinationBonus:
          type: number
...
    UpdateContainerChargesRequest:
      type: object
      properties:
        section:
          type: string
          enum: [shipping, customs, terminal, delivery, operations]
        shipping:
          $ref: "#/components/schemas/ShippingCharges"
        customs:
          $ref: "#/components/schemas/CustomsCharges"
        terminal:
          $ref: "#/components/schemas/TerminalCharges"
        delivery:
          $ref: "#/components/schemas/DeliveryCharges"
        operations:
          $ref: "#/components/schemas/OperationsCharges"
        clearingCharges:
          type: number
        reason:
          type: string
      required: [section]
    ContainerDetail:
      type: object
      properties:
        container:
          $ref: "#/components/schemas/Container"
        charges:
          $ref: "#/components/schemas/ContainerCharges"
      required: [container, charges]
    AuditEntry:
      type: object
      properties:
        id:
          type: integer
        containerId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        action:
          type: string
        section:
          type: string
          nullable: true
        fieldChanged:
          type: string
          nullable: true
        oldValue:
          type: string
          nullable: true
        newValue:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, containerId, userId, userName, action, createdAt]
...